Update the Tile Galleries block
