+++ client/gutenberg/extensions/tiled-gallery/index.js
@@ -1,0 +1,186 @@
+/** @format */
+
+/**
+ * External dependencies
+ */
+import { createBlock } from '@wordpress/blocks';
+import { filter } from 'lodash';
+import { Rect, SVG } from '@wordpress/components';
+
+/**
+ * Internal dependencies
+ */
+import { __, _x } from 'gutenberg/extensions/presets/jetpack/utils/i18n';
+import { default as edit } from './edit';
+import { default as save } from './save';
+import { DEFAULT_COLUMNS, DEFAULT_LAYOUT, LAYOUT_STYLES, LAYOUTS } from './constants';
+
+/**
+ * Style dependencies
+ */
+import './editor.scss';
+
+const blockAttributes = {
+	images: {
+		type: 'array',
+		default: [],
+		source: 'query',
+		selector: '.wp-block-jetpack-tiled-gallery .tiled-gallery__item',
+		query: {
+			url: {
+				source: 'attribute',
+				selector: 'img',
+				attribute: 'src',
+			},
+			link: {
+				source: 'attribute',
+				selector: 'img',
+				attribute: 'data-link',
+			},
+			alt: {
+				source: 'attribute',
+				selector: 'img',
+				attribute: 'alt',
+				default: '',
+			},
+			id: {
+				source: 'attribute',
+				selector: 'img',
+				attribute: 'data-id',
+			},
+			caption: {
+				source: 'html',
+				selector: 'figcaption',
+			},
+		},
+	},
+	columns: {
+		type: 'number',
+		default: DEFAULT_COLUMNS,
+	},
+	imageCrop: {
+		type: 'boolean',
+		default: true,
+	},
+	linkTo: {
+		type: 'string',
+		default: 'none',
+	},
+};
+
+export const name = 'tiled-gallery';
+
+export const settings = {
+	attributes: blockAttributes,
+	category: 'jetpack',
+	description: __( 'Display multiple images in an elegantly organized tiled layout.' ),
+	icon: (
+		<SVG viewBox="0 0 20 20">
+			<Rect x="8" y="11" width="9" height="6" />
+			<Rect x="3" y="11" width="4" height="6" />
+			<Rect x="13" y="7" width="4" height="3" />
+			<Rect x="13" y="3" width="4" height="3" />
+			<Rect x="3" y="3" width="9" height="7" />
+		</SVG>
+	),
+	keywords: [
+		_x( 'images', 'block search term' ),
+		_x( 'photos', 'block search term' ),
+		_x( 'masonry', 'block search term' ),
+	],
+	styles: LAYOUT_STYLES,
+	supports: {
+		align: true,
+		html: false,
+	},
+	title: __( 'Tiled gallery' ),
+	transforms: {
+		from: [
+			{
+				type: 'block',
+				blocks: [ 'core/gallery' ],
+				transform: attributes => {
+					const validImages = filter( attributes.images, ( { id, url } ) => id && url );
+					if ( validImages.length > 0 ) {
+						return createBlock( name, {
+							images: validImages.map( ( { id, url, alt, caption } ) => ( {
+								id,
+								url,
+								alt,
+								caption,
+							} ) ),
+						} );
+					}
+					return createBlock( name );
+				},
+			},
+			{
+				type: 'shortcode',
+				tag: 'gallery',
+				attributes: {
+					// @TODO: other params: https://en.support.wordpress.com/gallery/#gallery-shortcode
+					images: {
+						type: 'array',
+						shortcode: ( { named: { ids } } ) => {
+							if ( ! ids ) {
+								return [];
+							}
+
+							return ids.split( ',' ).map( id => ( {
+								id: parseInt( id, 10 ),
+							} ) );
+						},
+					},
+					columns: {
+						type: 'number',
+						shortcode: ( { named: { columns = 3 } } ) => {
+							if ( ! columns ) {
+								return;
+							}
+
+							const result = parseInt( columns, 10 );
+							if ( result ) {
+								return result;
+							}
+						},
+					},
+					linkTo: {
+						type: 'string',
+						shortcode: ( { named: { link = 'attachment' } } ) => {
+							return link === 'file' ? 'media' : link;
+						},
+					},
+					layout: {
+						type: 'string',
+						shortcode: ( { named: { type = DEFAULT_LAYOUT } } ) => {
+							// @TODO: if `type=slideshow`, return a slideshow block
+							return LAYOUTS.indexOf( type ) > -1 ? type : DEFAULT_LAYOUT;
+						},
+					},
+				},
+			},
+		],
+		to: [
+			{
+				type: 'block',
+				blocks: [ 'core/gallery' ],
+				transform: ( { images, columns, imageCrop, linkTo } ) =>
+					createBlock( 'core/gallery', { images, columns, imageCrop, linkTo } ),
+			},
+			{
+				type: 'block',
+				blocks: [ 'core/image' ],
+				transform: ( { images } ) => {
+					if ( images.length > 0 ) {
+						return images.map( ( { id, url, alt, caption } ) =>
+							createBlock( 'core/image', { id, url, alt, caption } )
+						);
+					}
+					return createBlock( 'core/image' );
+				},
+			},
+		],
+	},
+	edit,
+	save,
+};
